requirements.txt
google-genai
PyYAML
requests
tqdm
flask
openai
zhipuai
markitdown
pypinyin
Werkzeug